requirements.txt
django
djangorestframework
celery
requests
django-rest-passwordreset
psycopg2-binary
redis